requirements.txt
flask
python-telegram-bot
gspread
oauth2client
pandas
python-dotenv
google-generativeai
requests